Use nest IFixture impl for extension test
