+++ Src/AutoFixture.NUnit3.UnitTest/InlineAutoDataAttributeTest.cs
@@ -1,4 +1,8 @@
 ﻿using NUnit.Framework;
+using System;
+using Ploeh.AutoFixture.Dsl;
+using Ploeh.AutoFixture.Kernel;
+using System.Collections.Generic;
 
 namespace Ploeh.AutoFixture.NUnit3.UnitTest
 {
@@ -28,8 +32,81 @@ namespace Ploeh.AutoFixture.NUnit3.UnitTest
             /// </summary>
             /// <param name="arguments"></param>
             public InlineAutoDataAttributeStub(params object[] arguments) 
-                : base(new ThrowingStubFixture(), arguments)
+                : base(new StubFixture(), arguments)
             {
+            }
+        }
+
+        private class StubFixture : IFixture
+        {
+            public IList<ISpecimenBuilderTransformation> Behaviors
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            public IList<ISpecimenBuilder> Customizations
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            public bool OmitAutoProperties
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+
+                set
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            public int RepeatCount
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+
+                set
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            public IList<ISpecimenBuilder> ResidueCollectors
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+            }
+
+            public ICustomizationComposer<T> Build<T>()
+            {
+                throw new NotImplementedException();
+            }
+
+            public object Create(object request, ISpecimenContext context)
+            {
+                throw new NotImplementedException();
+            }
+
+            public IFixture Customize(ICustomization customization)
+            {
+                throw new NotImplementedException();
+            }
+
+            public void Customize<T>(Func<ICustomizationComposer<T>, ISpecimenBuilder> composerTransformation)
+            {
+                throw new NotImplementedException();
             }
         }
     }
